fix: Corregir errores en PosController y ZipFileDecompress   - Convertir valores a float antes de operaciones en validate_stock   - Verificar existencia de archivo temporal antes de eliminar en ZipFileDecompress

+++ app/CoreFacturalo/WS/Zip/ZipFileDecompress.php
@@ -25,7 +25,9 @@ class ZipFileDecompress
             $output = iterator_to_array($this->getFiles($zip, $filter));
         }
         $zip->close();
-        unlink($temp);
+        if (file_exists($temp)) {
+            unlink($temp);
+        }
 
         return $output;
     }
